Move `use_status_by_location` hook to main hooks directory and allow it to take arbitrary monitor IDs as an arg.
